Fix MSVC compiler warnings
- Fix C4129 unrecognized character escale sequence
- Fix C4305 truncation from double to float
- Fix C4244 conversion from double to float

+++ src/Math.hpp
@@ -37,14 +37,14 @@ using Vector3d = Vector3<double>;
 
 // standard-like functions
 inline float radf(float deg) {
-    return deg * M_PI / 180.0f;
+    return deg * static_cast<float>(M_PI) / 180.0f;
 }
 inline double rad(double deg) {
     return deg * M_PI / 180.0;
 }
 
 inline float degf(float rad) {
-    return rad * 180.0f / M_PI;
+    return rad * 180.0f / static_cast<float>(M_PI);
 }
 
 inline double deg(double rad) {
